requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==3.2.0
boto3==1.36.0
botocore==1.36.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.6
fitparse==1.2.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
jwt==1.3.1
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
s3transfer==0.11.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0